+++ pcweb/pages/docs/component_lib/forms/__init__.py
@@ -1,0 +1,11 @@
+from .forms import *
+from .button import *
+from .checkbox import *
+from .editable import *
+from .formcontrol import *
+from .input import *
+from .numberinput import *
+from .pininput import *
+from .radiogroup import *
+from .select import *
+from .slider import *
